Outputs mostly correct, pushing before coversion from cout to printf

+++ src/session.h
@@ -46,7 +46,7 @@ class Session {
 /**
  * Writes into a transaction file 
  */
-  void write_file(int trans_num, std::string name = "", int account_id = 0, float value = 0.0, std::string misc = "");
+  void write_file(int trans_num, std::string name = "", int account_id = 0, int value = 0, std::string misc = "");
 
 /**
  * Use to read standard input
